lf lint: lfd: move migration registry to dedicated module
